Get canvas hierarchical packed circles working.
- Make the same canvas render logic work for flat and hierarchical data.
- Move dummy data generators to a helper file.
